add form in Suppliers file

+++ src/api/articles.js
@@ -5,6 +5,10 @@ const app       = require("../app.js");
 app.route("/api/articles/create")
     .get((req, res) => res.status(503).send({ status: "ERROR"}))
     .post((req, res) => {
+        if (!req.user || !req.user.id) {
+            res.status(503).send({status: "ERROR", extra: "Log in please !"});
+            return;
+        }
         if (typeof req.body.title !== "string" || req.body.title === "") {
             res.status(503).send({ status: "ERROR", extra: "Title is required" });
             return;
@@ -13,7 +17,7 @@ app.route("/api/articles/create")
             res.status(503).send({ status: "ERROR", extra: "Content is empty" });
             return;
         }
-        if (typeof req.body.author !== "string" || req.body.author === "") {
+        if (typeof req.body.author !== "number" || req.body.author <= 0 ) {
             res.status(503).send({ status: "ERROR", extra: "Author is not registered" });
             return;
         }
@@ -37,13 +41,17 @@ app.route("/api/articles/create")
 app.route("/api/articles/delete")
     .get((req, res) => res.status(503).send({ status: "ERROR"}))
     .post((req, res) => {
+        if (!req.user || !req.user.id) {
+            res.status(503).send({status: "ERROR", extra: "Log in please !"});
+            return;
+        }
         if (typeof req.body.id !== "string" || req.body.id === "")  {
             res.status(503).send({ status: "ERROR", extra: "Article id is required" });
             return;
         } 
         const sqlConnection = mysql.createConnection(sqlConfig);
         sqlConnection.query(
-            "DELETE FROM node_articles WHERE id = ?;",
+            "DELETE FROM node_articles WHERE id = ? AND author = ?;",
             [ req.body.id ],
             (error, result) => {
                 if (error) {
@@ -52,7 +60,7 @@ app.route("/api/articles/delete")
                 } else {
                     console.log(result);
                     if (result.affectedRows === 0) {
-                        res.status(503).send({ status: "ERROR", extra: "Article doesn't exist" });
+                        res.status(503).send({ status: "ERROR", extra: "You cannnot delete this article" });
                         return;
                     } else {
                         res.send({ status: "OK" });
@@ -66,7 +74,7 @@ app.route("/api/articles/delete")
 app.get("/api/articles", (req, res) => {
     const sqlConnection = mysql.createConnection(sqlConfig);
     sqlConnection.query(
-        "SELECT node_articles.id, title, content, node_users.firstname AS authorFirstname, node_users.lastname AS authorLastname, created_at"
+        "SELECT node_articles.id, title, content, node_users.id AS authorId, node_users.firstname AS authorFirstname, node_users.lastname AS authorLastname, created_at"
         + "  FROM node_articles"
         + "  LEFT JOIN node_users"
         + "  ON node_articles.author = node_users.id"
@@ -90,7 +98,7 @@ app.get("/api/articles", (req, res) => {
 app.get("/api/article", (req, res) => {
     const sqlConnection = mysql.createConnection(sqlConfig);
     sqlConnection.query(
-        "SELECT node_articles.id, title, content, node_users.firstname AS authorFirstname, node_users.lastname AS authorLastname, created_at"
+        "SELECT node_articles.id, title, content, node_users.id AS authorId, node_users.firstname AS authorFirstname, node_users.lastname AS authorLastname, created_at"
         + "  FROM node_articles"
         + "  LEFT JOIN node_users"
         + "  ON node_articles.author = node_users.id"
